# src/gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Blog Posts Search Service
  description: HTTP service for blog posts search
  version: ""
host: localhost:8088
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
  /posts/bulk:
    post:
      tags:
      - posts
      summary: register posts
      description: registers blog posts to be searched
      operationId: posts#register
      parameters:
      - name: Authorization
        in: header
        description: API key used to perform authorization
        required: true
        type: string
      - name: RegisterRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/PostsRegisterRequestBody'
          required:
          - posts
      responses:
        "201":
          description: Created response.
          schema:
            type: integer
            format: int64
      schemes:
      - http
      security:
      - api_key_header_Authorization: []
  /posts/related:
    get:
      tags:
      - posts
      summary: relatedPosts posts
      description: get related blog posts
      operationId: posts#relatedPosts
      parameters:
      - name: url
        in: query
        description: post's url
        required: true
        type: string
      - name: count
        in: query
        description: count
        required: false
        type: integer
        default: 5
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/PostsRelatedPostsResponseBody'
            required:
            - posts
            - count
      schemes:
      - http
  /posts/search:
    get:
      tags:
      - posts
      summary: search posts
      description: search blog posts
      operationId: posts#search
      parameters:
      - name: query
        in: query
        description: search query
        required: true
        type: string
      - name: page
        in: query
        description: page
        required: false
        type: integer
        default: 1
      - name: pageSize
        in: query
        description: results per page
        required: false
        type: integer
        default: 50
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/PostsSearchResponseBody'
            required:
            - posts
            - page
            - totalPage
      schemes:
      - http
  /swagger.json:
    get:
      tags:
      - openapi
      summary: Download ../gen/http/openapi.json
      operationId: openapi#/swagger.json
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
      schemes:
      - http
definitions:
  PostOutputResponseBody:
    title: 'Mediatype identifier: application/vnd.posts; view=default'
    type: object
    properties:
      description:
        type: string
        description: Post's description
        example: Et recusandae cupiditate dolores voluptatibus quas vero.
      id:
        type: string
        description: Post's id
        example: Fugiat hic neque quo modi.
      screenImageUrl:
        type: string
        description: Post's screen image url
        example: Sit qui nesciunt in odio fugiat.
      title:
        type: string
        description: Post's title
        example: Non quaerat.
    description: PostOutputResponseBody result type (default view)
    example:
      description: Quae ipsa quaerat minus.
      id: Modi quibusdam temporibus quod facilis.
      screenImageUrl: Animi quam voluptatem et commodi maiores quidem.
      title: Iure incidunt occaecati.
    required:
    - id
    - title
    - description
    - screenImageUrl
  PostOutputResponseBodyCollection:
    title: 'Mediatype identifier: application/vnd.posts; type=collection; view=default'
    type: array
    items:
      $ref: '#/definitions/PostOutputResponseBody'
    description: PostOutputCollectionResponseBody is the result type for an array of PostOutputResponseBody (default view)
    example:
    - description: Enim architecto ea est.
      id: Officiis provident et excepturi.
      screenImageUrl: Nam neque.
      title: Iusto accusantium nemo.
    - description: Enim architecto ea est.
      id: Officiis provident et excepturi.
      screenImageUrl: Nam neque.
      title: Iusto accusantium nemo.
    - description: Enim architecto ea est.
      id: Officiis provident et excepturi.
      screenImageUrl: Nam neque.
      title: Iusto accusantium nemo.
  PostParamsRequestBody:
    title: PostParamsRequestBody
    type: object
    properties:
      body:
        type: string
        description: Post's body
        example: Quia suscipit aut.
      description:
        type: string
        description: Post's description
        example: Recusandae quia sit porro repellendus et non.
      id:
        type: string
        description: Post's id
        example: Unde autem est autem non.
      screenImageUrl:
        type: string
        description: Post's screen image url
        example: Nostrum nobis ratione rerum corrupti ipsam.
      title:
        type: string
        description: Post's title
        example: Necessitatibus ab.
    example:
      body: Labore quas sed.
      description: Ut sequi est qui omnis.
      id: Perferendis blanditiis.
      screenImageUrl: Non sunt architecto unde et ea facilis.
      title: Quo aliquam repudiandae et enim et.
  PostsRegisterRequestBody:
    title: PostsRegisterRequestBody
    type: object
    properties:
      posts:
        type: array
        items:
          $ref: '#/definitions/PostParamsRequestBody'
        example:
        - body: Harum molestiae nemo quo.
          description: Tenetur autem libero sint voluptatem.
          id: Est ipsa laboriosam assumenda veritatis sapiente ullam.
          screenImageUrl: Possimus illo voluptatibus corrupti.
          title: Debitis asperiores quasi.
        - body: Harum molestiae nemo quo.
          description: Tenetur autem libero sint voluptatem.
          id: Est ipsa laboriosam assumenda veritatis sapiente ullam.
          screenImageUrl: Possimus illo voluptatibus corrupti.
          title: Debitis asperiores quasi.
        minItems: 1
    example:
      posts:
      - body: Harum molestiae nemo quo.
        description: Tenetur autem libero sint voluptatem.
        id: Est ipsa laboriosam assumenda veritatis sapiente ullam.
        screenImageUrl: Possimus illo voluptatibus corrupti.
        title: Debitis asperiores quasi.
    required:
    - posts
  PostsRelatedPostsResponseBody:
    title: PostsRelatedPostsResponseBody
    type: object
    properties:
      count:
        type: integer
        example: 17815395042763388373
        format: int64
      posts:
        $ref: '#/definitions/PostOutputResponseBodyCollection'
    example:
      count: 4894117250138250688
      posts:
      - description: Assumenda nesciunt nesciunt quasi voluptates perferendis.
        id: Qui nihil.
        screenImageUrl: Tempore voluptas cumque voluptatem aut facere.
        title: Alias incidunt sunt ut veniam.
      - description: Assumenda nesciunt nesciunt quasi voluptates perferendis.
        id: Qui nihil.
        screenImageUrl: Tempore voluptas cumque voluptatem aut facere.
        title: Alias incidunt sunt ut veniam.
      - description: Assumenda nesciunt nesciunt quasi voluptates perferendis.
        id: Qui nihil.
        screenImageUrl: Tempore voluptas cumque voluptatem aut facere.
        title: Alias incidunt sunt ut veniam.
      - description: Assumenda nesciunt nesciunt quasi voluptates perferendis.
        id: Qui nihil.
        screenImageUrl: Tempore voluptas cumque voluptatem aut facere.
        title: Alias incidunt sunt ut veniam.
    required:
    - posts
    - count
  PostsSearchResponseBody:
    title: PostsSearchResponseBody
    type: object
    properties:
      page:
        type: integer
        example: 2215612457498736729
        format: int64
      posts:
        $ref: '#/definitions/PostOutputResponseBodyCollection'
      totalPage:
        type: integer
        example: 14151283819490208690
        format: int64
    example:
      page: 16573824922741616287
      posts:
      - description: Assumenda nesciunt nesciunt quasi voluptates perferendis.
        id: Qui nihil.
        screenImageUrl: Tempore voluptas cumque voluptatem aut facere.
        title: Alias incidunt sunt ut veniam.
      - description: Assumenda nesciunt nesciunt quasi voluptates perferendis.
        id: Qui nihil.
        screenImageUrl: Tempore voluptas cumque voluptatem aut facere.
        title: Alias incidunt sunt ut veniam.
      - description: Assumenda nesciunt nesciunt quasi voluptates perferendis.
        id: Qui nihil.
        screenImageUrl: Tempore voluptas cumque voluptatem aut facere.
        title: Alias incidunt sunt ut veniam.
      - description: Assumenda nesciunt nesciunt quasi voluptates perferendis.
        id: Qui nihil.
        screenImageUrl: Tempore voluptas cumque voluptatem aut facere.
        title: Alias incidunt sunt ut veniam.
      totalPage: 981018779888932699
    required:
    - posts
    - page
    - totalPage
securityDefinitions:
  api_key_header_Authorization:
    type: apiKey
    description: secret api key for authentication
    name: Authorization
    in: header